made autobytes4 and modified other autobytes files
autobytes4 creates an xml file that is the log of the scan. Then autobytes4 parses through the xml file to see if there are any threats that were removed or there were no threats. Then it sends an email that attaches the install file and scan file and tells if there were threats that existed and removed or if there were none.